Cleanup
Tries to unjankify things.
