+++ src/main/java/com/minecolonies/coremod/event/EventHandler.java
@@ -3,7 +3,6 @@ package com.minecolonies.coremod.event;
 import com.ldtteam.structures.helpers.Settings;
 import com.ldtteam.structurize.items.ModItems;
 import com.ldtteam.structurize.util.LanguageHandler;
-import com.minecolonies.api.IMinecoloniesAPI;
 import com.minecolonies.api.blocks.AbstractBlockHut;
 import com.minecolonies.api.blocks.interfaces.IRSComponentBlock;
 import com.minecolonies.api.colony.*;
@@ -12,11 +11,9 @@ import com.minecolonies.api.colony.buildings.IGuardBuilding;
 import com.minecolonies.api.colony.permissions.Action;
 import com.minecolonies.api.items.ItemBlockHut;
 import com.minecolonies.api.research.IGlobalResearchTree;
-import com.minecolonies.api.util.ItemStackUtils;
 import com.minecolonies.api.util.Log;
 import com.minecolonies.api.util.WorldUtil;
 import com.minecolonies.api.util.constant.Constants;
-import com.minecolonies.apiimp.initializer.ModTagsInitializer;
 import com.minecolonies.coremod.MineColonies;
 import com.minecolonies.coremod.Network;
 import com.minecolonies.coremod.blocks.BlockScarecrow;
@@ -37,7 +34,6 @@ import com.minecolonies.coremod.network.messages.client.UpdateChunkCapabilityMes
 import com.minecolonies.coremod.network.messages.client.UpdateChunkRangeCapabilityMessage;
 import com.minecolonies.coremod.util.ChunkClientDataHelper;
 import com.minecolonies.coremod.util.ChunkDataHelper;
-import com.minecolonies.coremod.util.FurnaceRecipes;
 import net.minecraft.block.*;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.entity.player.ClientPlayerEntity;
@@ -67,7 +63,6 @@ import net.minecraftforge.client.event.RenderGameOverlayEvent;
 import net.minecraftforge.event.AttachCapabilitiesEvent;
 import net.minecraftforge.event.LootTableLoadEvent;
 import net.minecraftforge.event.RegisterCommandsEvent;
-import net.minecraftforge.event.TagsUpdatedEvent;
 import net.minecraftforge.event.entity.EntityEvent;
 import net.minecraftforge.event.entity.EntityJoinWorldEvent;
 import net.minecraftforge.event.entity.EntityTravelToDimensionEvent;
@@ -763,22 +758,6 @@ public class EventHandler
             IColonyManager.getInstance().resetColonyViews();
             ItemBlockHut.checkResearch(null);
             Log.getLogger().info("Removed all colony views");
-        }
-    }
-
-    @SubscribeEvent
-    public static void onTagUpdate(final TagsUpdatedEvent.VanillaTagTypes event)
-    {
-        // This should only run on data pack reload or reconnection to a world; local single-player and dedicated server uses the FMLServerStartedEvent.
-        // This is also the only place where a Tag Supplier is guaranteed to have the output of a transmitted TagSupplier.
-        ModTagsInitializer.init(event.getTagManager());
-        IMinecoloniesAPI.getInstance().getColonyManager().getCompatibilityManager().invalidateTagsAndConfigs();
-        // Can't always run a discover here on first connection to any world, as the food and smelting-related recipes may not yet be filled.
-        // Those cases run in FurnaceRecipes.setMap (remote client) or FMLEventHandler.onServerStarted (logical server).
-        // Does run discover() run on reconnection, or data pack reloads.
-        if(ItemStackUtils.ISFOOD != null && FurnaceRecipes.instance.loaded())
-        {
-            IMinecoloniesAPI.getInstance().getColonyManager().getCompatibilityManager().discover();
         }
     }
 }
