+++ src/main/java/com/minecolonies/coremod/util/FurnaceRecipes.java
@@ -2,37 +2,25 @@ package com.minecolonies.coremod.util;
 
 import com.google.common.collect.ImmutableList;
 
-import com.minecolonies.api.IMinecoloniesAPI;
 import com.minecolonies.api.colony.requestsystem.StandardFactoryController;
 import com.minecolonies.api.compatibility.IFurnaceRecipes;
 import com.minecolonies.api.crafting.ItemStorage;
 import com.minecolonies.api.crafting.RecipeStorage;
 import com.minecolonies.api.util.ItemStackUtils;
-import com.minecolonies.api.util.constant.Constants;
 import com.minecolonies.api.util.constant.TypeConstants;
-import com.minecolonies.coremod.Network;
-import com.minecolonies.coremod.network.messages.client.UpdateClientWithRecipesMessage;
 import com.minecolonies.coremod.recipes.FoodIngredient;
 import net.minecraft.block.Blocks;
-import net.minecraft.entity.player.ServerPlayerEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.crafting.IRecipeType;
 import net.minecraft.item.crafting.Ingredient;
-import net.minecraft.server.MinecraftServer;
+import net.minecraft.item.crafting.RecipeManager;
 import net.minecraft.util.NonNullList;
-import net.minecraftforge.event.entity.player.PlayerEvent;
-import net.minecraftforge.eventbus.api.SubscribeEvent;
-import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.fml.event.server.FMLServerAboutToStartEvent;
-import net.minecraftforge.fml.event.server.FMLServerStartedEvent;
-import net.minecraftforge.registries.ObjectHolder;
+
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
 import java.util.function.Predicate;
 
-@ObjectHolder(Constants.MOD_ID)
-@Mod.EventBusSubscriber(modid = Constants.MOD_ID, bus = Mod.EventBusSubscriber.Bus.FORGE)
 public class FurnaceRecipes implements IFurnaceRecipes
 {
     /**
@@ -49,11 +37,14 @@ public class FurnaceRecipes implements IFurnaceRecipes
     /**
      * Load all the recipes in the recipe storage.
      *
-     * @param server the server obj to load.
+     * @param recipeManager  The recipe manager to parse.
      */
-    private void loadRecipes(final MinecraftServer server)
+    public void loadRecipes(final RecipeManager recipeManager)
     {
-        server.getRecipeManager().getRecipes(IRecipeType.SMELTING).values().forEach(recipe -> {
+        recipes.clear();
+        reverseRecipes.clear();
+        loadUtilityPredicates();
+        recipeManager.getRecipes(IRecipeType.SMELTING).values().forEach(recipe -> {
             final NonNullList<Ingredient> list = recipe.getIngredients();
             if (list.size() == 1)
             {
@@ -80,13 +71,6 @@ public class FurnaceRecipes implements IFurnaceRecipes
         });
     }
 
-    @SubscribeEvent
-    public static void onServerStarting(final FMLServerStartedEvent event)
-    {
-        instance.loadRecipes(event.getServer());
-        loadUtilityPredicates();
-    }
-
     /**
      * Load all the utility predicates.
      */
@@ -97,22 +81,6 @@ public class FurnaceRecipes implements IFurnaceRecipes
         ItemStackUtils.ISCOOKABLE = itemStack -> ItemStackUtils.ISFOOD.test(instance.getSmeltingResult(itemStack));
         ItemStackUtils.CAN_EAT =
                 itemStack -> ItemStackUtils.ISFOOD.test(itemStack) && !ItemStackUtils.ISCOOKABLE.test(itemStack);
-    }
-
-    @SubscribeEvent
-    public static void onServerStarting(final FMLServerAboutToStartEvent event)
-    {
-        instance = new FurnaceRecipes();
-        loadUtilityPredicates();
-    }
-
-    @SubscribeEvent
-    public static void onPlayerLogin(final PlayerEvent.PlayerLoggedInEvent event)
-    {
-        if (!event.getPlayer().world.isRemote)
-        {
-            Network.getNetwork().sendToPlayer(new UpdateClientWithRecipesMessage(instance.recipes), (ServerPlayerEntity) event.getPlayer());
-        }
     }
 
     /**
@@ -126,11 +94,6 @@ public class FurnaceRecipes implements IFurnaceRecipes
         if (ItemStackUtils.ISFOOD == null)
         {
             loadUtilityPredicates();
-        }
-        // This is the first place remote clients can reliably run discoverEdibles and discoverOres, as they depend on furnace recipes and predicates.
-        if(!IMinecoloniesAPI.getInstance().getColonyManager().getCompatibilityManager().isDiscoveredAlready())
-        {
-            IMinecoloniesAPI.getInstance().getColonyManager().getCompatibilityManager().discover();
         }
     }
